(feat) hasTableCell matches the string value of cell item
In some cases, it is useful to match a string value against a table
cell's value, regardless of type. For example, when testing a table
of doubles, the string comparison might allow for small discrepancies
in the digital representation that are imperceptible to the user.
This is consistent with BDD principles: implementation details aren't
relevant in testing.

In testfx-legacy, this functionality is provided in the
containsCell method. Without adding a suitable replacement in
testfx-core, it will be more difficult for some developers to
migrate to TestFX 4.