Add TextFlowMatchers used for matching the Text of a TextFlow.
Currently there are two ways to match the text contained in a TextFlow.
The first one is based simply off the contained text. The second is
based on the contained text *and* the color of that text. The expected
colors are encoded in the expected String like:

<COLOR>some text</COLOR>

where COLOR is one of the named colors in JavaFX. For a list of all such
colors, see:

https://docs.oracle.com/javafx/2/api/javafx/scene/doc-files/cssref.html#typecolor

We also extracted out a method for computing the distance between two
colors from PixelMatcherRgb as that functionality was also needed for
finding the color nearest to a named color.