Refactor ColorUtils to use a HashMap instead of an ArrayList.
Add a method to TextFlowMatchers that allows for testing that the colors
of a TextFlow are *exactly* the same. This usage encouraged us to refactor
the named colors ArrayList to a HashMap so that the ColorUtils function
that looks up exact matches is quite peformant.

To test the hasExactlyColoredText(String string) method it is sufficient
to test the failing case, which is demonstrated by creating a TextFlow
with the text "exact" with a fill set to rgb(51, 200, 50) which is one
red value away from the named color "LimeGreen" and then asserting that:

TextFlowMatchers.hasExactlyColoredText("<LIMEGREEN>exact</LIMEGREEN>");

fails, as expected.