(enhancement) Add module-info.java to testfx-core.
This is an attempt to begin the process of jigsaw modularization for
Java 9. Currently we are not taking advantage of modules for the
different parts of testfx-core, we are simply making everything one big
module but it's the first step.

hamcrest-core and hamcrest-library were used in the compileTestJava
phase of the build and these two artifacts have a split package (both
use the same org.hamcrest package). Instead of messing about with
--patch-modules and the like we replaced all usages of hamcrest-library
with hamcrest-core CoreMatchers where possible. The only place where
this was not feasible was the "closeTo" matcher, which we copied into
CaptureSupportImplTest wholesale. If it turns out we need this Matcher
across the tests we will use a different method of "fixing" this.