(enhancement) Produce better descriptions of actual if mismatch occurs.
Add enhanced mismatch descriptions for ComboBoxMatchers,
LabeledMatchers, ListViewMatchers, TableViewMatchers, TextFlowMatchers,
TextInputControlMatchers and TextMatchers.

Also change matchers for some control to return an instance of
Matcher<Control> instead of Matcher`, this way type mistmatches are
caught at compile time instead of run-time:

Before this PR the following test would pass:

    private Region region;

    @Before
    public void setup() throws Exception {
        FxToolkit.setupFixture(() -> {
            region = new Region();
        });
    }

    @Test
    public void hasText_matcher_with_region_fails() {
        // expect:
        exception.expect(AssertionError.class);
        exception.expectMessage("Expected: Labeled has a string ending with \"bar\"\n");

        assertThat(region, LabeledMatchers.hasText(endsWith("bar")));
    }

This assertion will always fail (and thus the test itself is correct)
but somewhat pointless. After this PR trying to compile the line:

assertThat(region, LabeledMatchers.hasText(endsWith("bar")));

will result in an error because Region is not an instance of Labeled.

This is a side-effect of the fact that we need the type to create
"enhanced" mismatch descriptions but I think it may be better
regardless.

Replace deprecated mockito Matchers with ArgumentMatchers.
Update Javadocs and remove some boilerplate.
Replace multiple @Rule's with RuleChain's.
Fix TestFXRule implementation so that tests actually run.