(fix) Invoke ParametersImpl directly.
Using reflection gains us nothing and in fact obfuscates the fact that
we require the exporting of javafx.graphics/com.sun.javafx.graphics to
the org.testfx module.

This is also done for some of the Robot classes. The only way to get rid
of these explicit exports is if the JavaFX developers move them to the
public API.