Switch default robot to glass on all platforms.
This will break Appveyor builds that are using AWT robot with HiDPI (by
setting "glass.win.uiScale" to a value above 100%) because, for some
unknown reason as of yet, Appveyor behaves differently with respect to
display scaling (HiDPI) than real, local Windows machines. Therefore we
add the documentation about using Appveyor (with the new
"testfx.awt.scale" property set accordingly) to the README so it is more
visible.

Note for future work: Test running TestFX tests on Windows Server 2012+
(the OS Appveyor uses) with AWT robot and HiDPI on a local machine
(connected to a real display) and see if it behaves differently than
Appveyor.

Fixes #581